Added a command parser
When a command is entered in the console, the JTable is updated to
reflect the cells specified within the command. Keep in mind that to
view the table you have to write "exit" (without the quotes) in the
console. This will close the Scanner object and then display the updated
table. I have yet to clean up the code (take out unused variables, group
a list of lines within a method, etc.) I'll try to do it when I can. I
added some comments and enclosed my code within 2 lines of comments.
Obviously the parser isn't error proof so if you encounter any
particular errors let me know.