Merged my code and Max's, added name to .tex file
I changed the size of the 2D arrays from [10][11] to [10][10], the table
column now go from A-J and not A-K because we have to use a 10x10 table
and not 10x11. I also added the updateTable method which checks for
changes in the table. For example if the value of cell A1  is changed
all other cells that have  A1 in their formula will be updated
accordingly. I added a getCellName method where we can find the name of
a cell by entering its index i.e. getCellName(0,0) returns A1. I also
added my name to HelpDocument.tex